Small fixes

+++ Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: winorth <winorth@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/12 11:45:51 by winorth           #+#    #+#             */
-/*   Updated: 2019/03/04 15:37:16 by winorth          ###   ########.fr       */
+/*   Updated: 2019/03/13 20:09:48 by winorth          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int i;
